-Changed Back PID to return float Will increase precision on motor rotation when Yaw + Pitch/roll will be in use

+++ PILOT/pid.cpp
@@ -40,7 +40,7 @@ PID::PID()
   m_outmin = -20;
 }
 
-int PID::update_pid(float setpoint, float input, float dt)
+float PID::update_pid(float setpoint, float input, float dt)
 {
 
   //Computes error
@@ -72,7 +72,7 @@ int PID::update_pid(float setpoint, float input, float dt)
   //printf("setpt %f input %f outpt %f\n", m_Kp, m_Ki, m_Kd);
   //printf("setpt %f input %f outpt %f\n", setpoint, input, output);
 
-  return (int)m_output;
+  return m_output;
 }
 
 void PID::set_Kpid(float Kp,float Ki, float Kd)
